third

+++ app/api/employees/[id]/route.js
@@ -1,7 +1,26 @@
 import prisma from "@utils/connection";
+export async function GET(req, { params }) {
+    console.log(params)
+    const id = params.id;
 
-export async function PUT(req) {
-    const { id } = req.query; // Ambil ID dari URL
+    if (!id) {
+        return new Response('Employee ID is required', { status: 400 });
+    }
+
+    const employee = await prisma.employees.findUnique({
+        where: { id: id },
+    });
+
+    if (!employee) {
+        return new Response('Employee not found', { status: 404 });
+    }
+
+    return new Response(JSON.stringify(employee), {
+        headers: { 'Content-Type': 'application/json' },
+    });
+}
+export async function PUT(req, { params }) {
+    const id = params.id;
     const { name } = await req.json();
 
     if (!id) {
@@ -10,7 +29,7 @@ export async function PUT(req) {
 
     const updatedEmployee = await prisma.employees.update({
         where: { id },
-        data: { name }, // Tambahkan data lain yang ingin diupdate
+        data: { name },
     });
 
     return new Response(JSON.stringify(updatedEmployee), {
@@ -18,11 +37,22 @@ export async function PUT(req) {
     });
 }
 
-export async function DELETE(req) {
-    const { id } = req.query; // Ambil ID dari URL
+export async function DELETE(req, { params }) {
+    const id = params.id;
 
     if (!id) {
         return new Response('Employee ID is required', { status: 400 });
+    }
+
+    const checkWorkingHours = await prisma.workinghours.findMany({
+        where: { employeeId: id },
+    });
+
+    if (checkWorkingHours.length > 0) {
+        return new Response(JSON.stringify({ message: 'Karyawan masih memiliki jam kerja.' }), {
+            status: 400, // Status error
+            headers: { 'Content-Type': 'application/json' }
+        });
     }
 
     await prisma.employees.delete({
