+++ app/api/projects/[id]/route.js
@@ -1,8 +1,29 @@
 import prisma from "@utils/connection";
 
-export async function PUT(req) {
-    const { id } = req.query; // Ambil ID dari URL
-    const { name } = await req.json();
+export async function GET(req, { params }) {
+    console.log(params)
+    const id = params.id;
+
+    if (!id) {
+        return new Response('Project ID is required', { status: 400 });
+    }
+
+    const project = await prisma.project.findUnique({
+        where: { id: id },
+    });
+
+    if (!project) {
+        return new Response('Project not found', { status: 404 });
+    }
+
+    return new Response(JSON.stringify(project), {
+        headers: { 'Content-Type': 'application/json' },
+    });
+}
+
+export async function PUT(req, { params }) {
+    const id = params.id;
+    const { name, location, description } = await req.json();
 
     if (!id) {
         return new Response('Project ID is required', { status: 400 });
@@ -10,7 +31,7 @@ export async function PUT(req) {
 
     const updatedProject = await prisma.project.update({
         where: { id },
-        data: { name }, // Tambahkan data lain yang ingin diupdate
+        data: { name, location, description },
     });
 
     return new Response(JSON.stringify(updatedProject), {
@@ -18,11 +39,24 @@ export async function PUT(req) {
     });
 }
 
-export async function DELETE(req) {
-    const { id } = req.query; // Ambil ID dari URL
+export async function DELETE(req, { params }) {
+    const id = params.id;
 
     if (!id) {
         return new Response('Project ID is required', { status: 400 });
+    }
+
+
+
+    const checkWorkingHours = await prisma.workinghours.findMany({
+        where: { projectId: id },
+    });
+
+    if (checkWorkingHours.length > 0) {
+        return new Response(JSON.stringify({ message: 'Project masih memiliki riwayat jam kerja.' }), {
+            status: 400, // Status error
+            headers: { 'Content-Type': 'application/json' }
+        });
     }
 
     await prisma.project.delete({
