+++ app/employees/page.js
@@ -1,28 +1,29 @@
-// employees.js (Client Component)
-"use client"; // Menandakan ini adalah komponen client
+"use client";
 
+import useAlert from '@components/alert';
 import { useEffect, useState } from 'react';
 import { useRouter } from 'next/navigation';
 import Provider from "@components/provider";
 import Nav from "@components/nav";
 
 const employees = () => {
+    const { alert, showAlert, alertClass } = useAlert();
     const [name, setName] = useState('');
     const router = useRouter();
 
     const [employee, setEmployee] = useState([]);
 
-    useEffect(() => {
-        const fetchEmployee = async () => {
-            const response = await fetch('/api/employees/get');
-            if (!response.ok) {
-                console.error('Error fetching data:', response.statusText); // Log status error
-                return; // Keluar dari fungsi jika ada kesalahan
-            }
-            const data = await response.json();
-            setEmployee(data);
-        };
+    const fetchEmployee = async () => {
+        const response = await fetch('/api/employees/get');
+        if (!response.ok) {
+            console.error('Error fetching data:', response.statusText);
+            return;
+        }
+        const data = await response.json();
+        setEmployee(data);
+    };
 
+    useEffect(() => {
         fetchEmployee();
     }, []);
 
@@ -39,56 +40,49 @@ const employees = () => {
         });
 
         if (response.ok) {
-            // Jika berhasil, redirect ke halaman proyek
-
-            router.push('/employees');
+            showAlert('Data berhasil ditambahkan!', 'success');
+            window.scrollTo({ top: 0, behavior: 'smooth' });
+            fetchEmployee();
         } else {
             const errorData = await response.json();
             console.error('Error adding employee:', errorData.error);
         }
     };
 
-    const updateEmployee = async (id, name) => {
-        const response = await fetch(`/api/employees/${id}`, {
-            method: 'PUT',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ id, name }),
-        });
-
-        const data = await response.json();
-        if (response.ok) {
-            console.log('Employee updated:', data);
-        } else {
-            console.error('Error updating employee:', data);
-        }
-    };
-
-
     const deleteEmployee = async (id) => {
-        const response = await fetch(`/api/employees/${id}`, {
-            method: 'DELETE',
-            headers: {
-                'Content-Type': 'application/json',
-            },
-            body: JSON.stringify({ id }),
-        });
+        const confirmed = window.confirm("Apakah Anda yakin ingin menghapus karyawan ini?");
+        if (confirmed) {
+            const response = await fetch(`/api/employees/${id}`, {
+                method: 'DELETE',
+                headers: {
+                    'Content-Type': 'application/json',
+                },
+                body: JSON.stringify({ id }),
+            });
 
-        const data = await response.json();
-        if (response.ok) {
-            console.log(data.message);
-        } else {
-            console.error('Error deleting employee:', data);
+            const data = await response.json();
+            if (response.ok) {
+                showAlert('Data berhasil dihapus!', 'success');
+                fetchEmployee();
+                console.log(data.message);
+            } else {
+                showAlert(data.message, 'error');
+                window.scrollTo({ top: 0, behavior: 'smooth' });
+                console.error('Error deleting employee:', data);
+            }
         }
     };
-
 
 
     return (
         <Provider>
             <Nav />
-            <div className="w-10/12 mx-auto mt-48">
+            <div className="w-10/12 mx-auto my-48">
+                {alert.visible && (
+                    <div className={`p-4 mb-4 ${alertClass()} bg-${alert.type == 'success' ? 'green' : alert.type == 'error' ? 'red' : 'blue'}-100 rounded`} role="alert">
+                        {alert.message}
+                    </div>
+                )}
                 <div className="container mx-auto p-4">
                     <h1 className="text-2xl font-bold mb-4">Add New Employee</h1>
                     <form onSubmit={handleSubmit} className="space-y-4">
@@ -132,7 +126,7 @@ const employees = () => {
                                         {new Date(p.createdAt).toLocaleDateString()}
                                     </td>
                                     <td className="px-6 py-4">
-                                        <a href="#" className="font-medium text-blue-600 dark:text-blue-500 hover:underline">Edit</a>
+                                        <a href={`/employees/${p.id}`} className="font-medium text-blue-600 dark:text-blue-500 hover:underline">Edit</a>
                                         <button type='button' onClick={() => deleteEmployee(p.id)} className="font-medium ms-5 text-red-600 dark:text-red-500 hover:underline">Delete</button>
                                     </td>
                                 </tr>
